Refactor matrix generation functions
